Added subscripted setters
Enhanced subscript operators, providing setters for both .Map and
.Array style MessagePackValues.

Subscript operators now always return a MessagePackValue, if only just
a .Nil, allowing if-let-less traversal of nested MessagePackValue
enumerations.

Count now always returns a value  - 0 for non .Map or .Array types.
This provides the start of a path to adding CollectionType conformance
to the enumeration.